refactor:useMutation naming & pages folder naming
