+++ client/src/router/Router.tsx
@@ -7,12 +7,12 @@ import Exam from '../pages/Exam';
 import ExamCreate from '../pages/ExamCreate';
 import FindId from '../pages/FindId/FindId';
 import FindPw from '../pages/findPw/FindPw';
-import Home from '../pages/home/Home';
+import Home from '../pages/home';
 import Login from '../pages/Login';
 import MyPage from '../pages/MyPage';
 import PageNotFound from '../pages/PageNotFound';
-import Workbook from '../pages/workbook/Workbook';
-import WorkbookCreate from '../pages/workbookCreate/WorkbookCreate';
+import Workbook from '../pages/workbook';
+import WorkbookCreate from '../pages/workbookCreate';
 import Review from '../pages/Review';
 import Search from '../pages/Search';
 import SignUp from '../pages/signup/SignUp';
